VLM_Client/requirements.txt
# Dependencias del proyecto VLM_Client
# Instalar con: pip install -r requirements.txt

langchain>=0.3.0
langchain-openai>=0.3.0
langchain-core>=0.3.0
langgraph>=1.0.0
langgraph-prebuilt>=1.0.0

openai>=1.0.0

Pillow>=10.0.0
numpy>=1.24.0

requests>=2.31.0
python-dotenv>=1.0.0

opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp>=1.20.0
